Remove the observer method
The observer is based on `setInterval` which is not ideal. The idea is that I'll give the developer the hand to choose what kind of DOM change detection he wants.
* This has two advantages, less dependencies, code and useless polling with a `setInterval`.
* The developer has full control over when he wants to trigger an SVG redraw.
As observing a DOM Resize/move event still is not broadly supported on browsers, this will be the safe bet for now. I'll provide a demo with a resize/move event detector I've found here soon:
http://www.backalleycoder.com/2013/03/18/cross-browser-event-based-element-resize-detection/